Added Some new Extra Features
I added a text file that contain the acts like a database for the bank system. I also replaced the vector of clients with a map that reads the file and I made the key of the map is the account ID and the value is an object of type Client. I also added destructor for the BankingApplication class that is called when the employee want to exit the application it will automatically overrides the existing data with the new data + previous data.